requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
alembic
python-jose
pytest